fix actuator endpoints
