+++ core/cas-server-core-web/src/main/java/org/apereo/cas/config/CasCoreWebConfiguration.java
@@ -198,90 +198,4 @@ class CasCoreWebConfiguration {
             };
         }
     }
-
-    @Configuration(value = "CasCoreWebEndpointsConfiguration", proxyBeanMethods = false)
-    @EnableConfigurationProperties({CasConfigurationProperties.class, CorsEndpointProperties.class, WebProperties.class})
-    static class CasCoreWebEndpointsConfiguration {
-        @Bean
-        @RefreshScope(proxyMode = ScopedProxyMode.DEFAULT)
-        @ConditionalOnMissingBean(name = "casProtocolEndpointConfigurer")
-        public CasWebSecurityConfigurer<Void> casProtocolEndpointConfigurer() {
-            return new CasWebSecurityConfigurer<>() {
-                @Override
-                public List<String> getIgnoredEndpoints() {
-                    return List.of(
-                        StringUtils.prependIfMissing(CasProtocolConstants.ENDPOINT_LOGIN, "/"),
-                        StringUtils.prependIfMissing(CasProtocolConstants.ENDPOINT_LOGOUT, "/"),
-                        StringUtils.prependIfMissing(CasProtocolConstants.ENDPOINT_VALIDATE, "/"),
-                        StringUtils.prependIfMissing(CasProtocolConstants.ENDPOINT_SERVICE_VALIDATE, "/"),
-                        StringUtils.prependIfMissing(CasProtocolConstants.ENDPOINT_SERVICE_VALIDATE_V3, "/"),
-                        StringUtils.prependIfMissing(CasProtocolConstants.ENDPOINT_PROXY_VALIDATE, "/"),
-                        StringUtils.prependIfMissing(CasProtocolConstants.ENDPOINT_PROXY_VALIDATE_V3, "/"),
-                        StringUtils.prependIfMissing(CasProtocolConstants.ENDPOINT_PROXY, "/"));
-                }
-            };
-        }
-
-        @Bean
-        @ConditionalOnMissingBean(name = "restControllerEndpointDiscoverer")
-        public RestActuatorEndpointDiscoverer restControllerEndpointDiscoverer(
-            final ConfigurableApplicationContext applicationContext,
-            final ObjectProvider<PathMapper> endpointPathMappers,
-            final ObjectProvider<Collection<EndpointFilter<RestActuatorControllerEndpoint>>> filters) {
-            return new RestActuatorEndpointDiscoverer(applicationContext,
-                endpointPathMappers.orderedStream().toList(),
-                filters.getIfAvailable(Collections::emptyList));
-        }
-
-        @Bean
-        public WebMvcEndpointHandlerMapping webEndpointServletHandlerMapping(
-            final EndpointLinksResolver endpointLinksResolver,
-            final WebEndpointsSupplier webEndpointsSupplier,
-            final EndpointsSupplier<RestActuatorControllerEndpoint> restEndpointsSupplier,
-            final EndpointMediaTypes endpointMediaTypes,
-            final CorsEndpointProperties corsProperties,
-            final WebEndpointProperties webEndpointProperties,
-            final Environment environment) {
-            val webEndpoints = webEndpointsSupplier.getEndpoints();
-            val basePath = webEndpointProperties.getBasePath();
-            val endpointMapping = new EndpointMapping(basePath);
-            val shouldRegisterLinksMapping = shouldRegisterLinksMapping(webEndpointProperties, environment, basePath);
-            return new WebMvcEndpointHandlerMapping(endpointMapping, webEndpoints, endpointMediaTypes,
-                corsProperties.toCorsConfiguration(), endpointLinksResolver, shouldRegisterLinksMapping);
-        }
-
-        @Bean
-        public EndpointLinksResolver endpointLinksResolver(
-            final WebEndpointsSupplier webEndpointsSupplier,
-            final EndpointsSupplier<RestActuatorControllerEndpoint> restEndpointsSupplier,
-            final EndpointMediaTypes endpointMediaTypes,
-            final CorsEndpointProperties corsProperties,
-            final WebEndpointProperties webEndpointProperties) {
-            val basePath = webEndpointProperties.getBasePath();
-            val allEndpoints = new ArrayList<ExposableEndpoint<?>>();
-            val webEndpoints = webEndpointsSupplier.getEndpoints();
-            allEndpoints.addAll(webEndpoints);
-            allEndpoints.addAll(restEndpointsSupplier.getEndpoints());
-            return new EndpointLinksResolver(allEndpoints, basePath);
-        }
-
-        @Bean
-        @ConditionalOnMissingBean(name = "restControllerEndpointHandlerMapping")
-        public RestActuatorEndpointHandlerMapping restControllerEndpointHandlerMapping(
-            final EndpointsSupplier<RestActuatorControllerEndpoint> restEndpointsSupplier,
-            final CorsEndpointProperties corsProperties, final WebEndpointProperties webEndpointProperties) {
-            val endpointMapping = new EndpointMapping(webEndpointProperties.getBasePath());
-            return new RestActuatorEndpointHandlerMapping(endpointMapping, restEndpointsSupplier.getEndpoints(),
-                corsProperties.toCorsConfiguration());
-        }
-
-        private static boolean shouldRegisterLinksMapping(
-            final WebEndpointProperties webEndpointProperties,
-            final Environment environment,
-            final String basePath) {
-            return webEndpointProperties.getDiscovery().isEnabled()
-                && (org.springframework.util.StringUtils.hasText(basePath)
-                || ManagementPortType.get(environment) == ManagementPortType.DIFFERENT);
-        }
-    }
 }
