+++ core/cas-server-core-web/src/main/java/org/apereo/cas/config/CasCoreWebEndpointsAutoConfiguration.java
@@ -1,0 +1,21 @@
+package org.apereo.cas.config;
+
+import org.apereo.cas.configuration.CasConfigurationProperties;
+import org.apereo.cas.configuration.features.CasFeatureModule;
+import org.apereo.cas.util.spring.boot.ConditionalOnFeatureEnabled;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Import;
+
+/**
+ * This is {@link CasCoreWebAutoConfiguration}.
+ *
+ * @author Janis Semper
+ * @since 7.2.0
+ */
+@EnableConfigurationProperties(CasConfigurationProperties.class)
+@ConditionalOnFeatureEnabled(feature = CasFeatureModule.FeatureCatalog.Core)
+@AutoConfiguration
+@Import({CasCoreWebEndpointsConfiguration.class})
+public class CasCoreWebEndpointsAutoConfiguration {
+}
