+++ core/cas-server-core-web/src/main/resources/META-INF/spring/org.springframework.boot.actuate.autoconfigure.web.ManagementContextConfiguration.imports
@@ -1,0 +1,1 @@
+org.apereo.cas.config.CasCoreWebEndpointsAutoConfiguration
